requirement.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
Flask==2.0.3 Werkzeug==2.0.3
python-dotenv
pandas
osmnx